requirements.txt
django>=4.1.1,<4.2.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow